fix the null alias bug
In the administrator interface, people can save a dynamic service with no value for alias (I am sure that should be a bug too). 
That will make all the other dynamic services inaccessible. All the requests will return "the layer XXX is not found". 
This fix will solve that problem.